requirements.txt
langchain
langchain-openai
langchain-community
langgraph
streamlit
tavily-python
python-dotenv
chromadb
tiktoken
beautifulsoup4
